fix: avoid strings with no common subsequences
The diff algorithm from Wu et al. assumes that the two strings being
diff'd have at least one common subsequence. If they do not, the
algorithm fails, as the outcome is nondeterministic. To avoid this, I
added logic that detects if there are no common subsequences and
generates an arbitrary edit list that performs the correct
transformation.