chore: added helper for checking diffable values
While its useful to properly identify the type of diffable value we
are working with, it is also handy to know when a random value we
have is diffable. This helper prevents me from writing long
conditions in if statements checking if something is a string, an
object, or an array.