fix: diff ignores functions in objects on deletion
For our purposes, if a function is deleted from an object, we want to
ignore that. This is due to functions not being serializable and
never entering into the Yjs shared types. Because of this, every
update coming from Yjs to Zustand will be missing functions, which
the change list will tell us to delete if we don't ignore them.